refactor: 클래스명 수정 - Ladder -> Lines - LadderGame -> Ladder
