+++ src/main/java/nextstep/ladder/Main.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.LadderGame;
+import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
 
@@ -9,7 +9,7 @@ public class Main {
         String names = InputView.inputNames();
         int height = InputView.inputLadderHeight();
 
-        LadderGame ladderGame = new LadderGame(names, height);
+        Ladder ladderGame = new Ladder(names, height);
 
         ResultView.printResult(ladderGame);
     }
